Rename secret/remove to secret/delete
